+++ src/resolution/frameworks/index.ts
@@ -25,6 +25,7 @@ import { angularResolver } from './angular';
 import { analogResolver } from './analog';
 import { solidRouterResolver } from './solid-router';
 import { solidStartResolver } from './solid-start';
+import { qwikCityResolver } from './qwik-city';
 import { djangoResolver, flaskResolver, fastapiResolver } from './python';
 import { railsResolver } from './ruby';
 import { springResolver } from './java';
@@ -73,6 +74,7 @@ const FRAMEWORK_RESOLVERS: FrameworkResolver[] = [
   analogResolver,
   solidRouterResolver,
   solidStartResolver,
+  qwikCityResolver,
   // Python
   djangoResolver,
   flaskResolver,